Implement full e*l optimization
This patch implements the full e*l. This is a full dakka optimization
that will just try to scan forward over all preceding expressions to
a literal if it can. This means we have to try all the possible split
points, which can take a minute. It's some kind of exponential in the
worst case, but we mostly don't hit that.