Experiement with having multiple accepting states.
I ran the problem I was having with (^a)+ past Nate Bragg,
and he pointed out that I might be able to solve it by
moving towards a more typical representation of NFAs
which makes matching an annotation on a state rather
than a special type of state. This branch represents me
playing with that idea. Currently I'm not sure that it is
worth it because it adds some extra bookkeeping to the
inner loop to track the previous state, and intermediate
states make figuring out what that previous state ought
to be tricky. I don't want to rule the idea out, but I'm
giving up on it for now.